requirements.txt
fastapi
uvicorn
pillow
pytesseract
python-multipart
jinja2
requests
aiofiles
python-dotenv